+++ SBO-Lib/lib/SBO/Lib/Help.pm
@@ -13,6 +13,7 @@ use Exporter 'import';
 our @EXPORT_OK = qw{
   @help_batch
   @help_builds
+  @help_builds_fail_menu
   @help_clean
   @help_hints
   @help_list_mgmt
@@ -65,6 +66,14 @@ our @help_builds = ('Build Lists - Help',
 the "Filter" button to search within the results.
 
 "Filter" appears only for lists of more than one SlackBuild.');
+
+our @help_builds_fail_menu = ('Build Lists - Help',
+
+'Select a SlackBuild from this screen to see its "Operations" menu. Use
+the "Reinst.All" button to reinstall all packages on the list.
+
+The "Filter" button does not appear on lists related to package test
+failures.');
 
 our @help_clean = ('sboclean - Help',