+++ SBO-Lib/ChangeLog.md
@@ -4,6 +4,7 @@
   * Drop support for Slackware 14.0, 14.1 and 14.2
   * sbocheck, sbotool: Add a perl package compatibility test
   * sbocheck: --obsolete-check also downloads the perl version history file
+  * sbotool: Add a reinstallation button for build test failure lists
 
 #### 4.0.2 - 2025-10-31
   * sbotool: Package test fixes